Remove @user from UserRegistrationsController
Previously the 'new' and 'create' actions would assing the @user var for
the 'new' template to use.

This change removes that and goes back to how devise does it, which is just
`respond_with resource`. This allows us to remove the override for the 'new'
action which just assigns @user.