Call set_current_order explicitly where we need it
set_current_order was previously called in a before_filter in the
ControllerHelpers::Order module.  This is hitting the database on just
about every frontend page load when we really just need it when someone
is signing in.  This moves that call to signing in and will be
accompanied with an explanation for people that don't use solidus auth
devise.