Allow Spree::Admin::UserPasswordsController to be accessed from admin panel.
Since the Spree::Admin::UserPasswordsController was not being used anywhere within solidus the routes were not working. In order to get it working I needed to add a custom route in routes.rb because the default one was being overridden by the frontend route. Since we are calling this endpoint from the admin, we need to skip the action require_no_authentication because we are already authorizing the admin elsewhere. Without it we receive the error "Filter chain halted as :require_no_authentication rendered or redirected"