Pull highline out of user seed file
Highline isn't a dependency of this project. Also we don't use any
complicated functionality of highline so its pretty simple to switch to
puts and gets. This will make your password echo to the terminal, but
we don't recommend this should be used in prod so that shouldn't be a
problem.