Stop raising exception for undefined callback
The specs against Solidus master are failing because `set_current_order`
is not defined as a process_action callback. This used to be the case
but Solidus has removed this and slated it for v2.4.

For compatibility with older versions of Solidus, we should not raise an
ArgumentError if this callback is not defined but still skip it if it
is.

PR removing this callback:
https://github.com/solidusio/solidus/pull/2185