Test that we are acting as paranoid not it's behaviour
We use `acts_as_paranoid` to enable soft delete. The current spec
for `destroy` tests the behaviour of the `acts_as_paranoid` gem.

Also it tests behaviour that is not possible with latest version of
Solidus (86d0e6281cf04a856766f7da1506d8007164f0ae) any more.
We can't destroy users that have orders assigned to them.

This is sufficienlty tested in Solidus core and must not be tested here again.
Therefore I think it is safe to remove this spec and replace it with a basic
duck typing test to ensure we are still using `acts_as_paranoid`.