Don't reuse spree_current_user as `@user`
Although `@user` and `spree_current_user` refer to the same record on the DB,
they must be kept two different Ruby objects in order to avoid annoyances when
the user update process fails.

For example, consider a custom user edit page that shows both the user email in
the header and the edit form allows email customization.

When submitting the form with an invalid email, this email will be redisplayed
both in the header and in the form.