Add DB index to `reset_password_token` field
This field should have an index in order to ensure uniqueness also at DB level
and to speed up search performances.

The index is added only if it does not exist yet and if the field actually
exists on the DB.

This index is present also in DB migrations generated by Devise gem.

In order to be nice to installations that may have already added the index with
the standard Rails/Devise name, the migration uses a custom name. This helps
avoiding issues when rolling back the migration.