Fix tests for Devise's :confirmable module
The `set_confirmable_options` helper was fundamentally broken: the
code being tested is only loaded once, so simply changing values in
Spree::Auth::Config does not have any effect on the modules loaded
in Spree::User.

The only way to ensure we load the right modules is removing and
reloading the Spree::User class after stubbing the configuration.