Remove admin orders controller specs
These authentication concerns are already covered by specs
in Core. This may have sense in the past, when we have a
decorator in place for that controller, but it has been
removed with 1a0f531400bf.

One of these specs was also failing due to some recent change
in core solidusio/solidus#3728, showing the fragility of the
approach.