Decorate `Spree::UsersController` when already existing
This commit allows to decorate the class `Spree::UsersController`
when already present (ie. when it's already included in the gem
`solidus_frontend`) or to define the class from scratch when it's
not (ie. when using a version of Solidus where `solidus_frontend`
does not include it).

See solidusio/solidus#2695 for the PR that adds the account page
to Solidus Frontend.

Once all supported versions of Solidus include the account page,
we will be able to get rid of the class definition, together with
the ERB view at `lib/views/frontend/spree/users/show.html.erb`.