Fixes infinite redirection bug on password change
When signout after password change was enabled, request to update user password was resulting in redirection to account edit page. Since user was signedout, this was ending up in infinite redirection (due to preveious page was being account edit page).
As a fix, changed the redirection in above case to login page as user is already signed out. Added controller test to assert this case.