Update CI configuration
Run different Solidus versions on different jobs for better error handling.

We're now explicitly providing the ruby version to the executor while using a matrix configuration for better extensibility.

We're now running the linter on Ruby v3.0, as that's the minimal
requirement for Solidus master. As a consequence, we disable the cop for
following the gemspec version.