Set User Class on Boot
For most of the Rails boot process, an app running solidus_auth_devise
believes that its user class is whatever is configured in the
config/spree.rb initializer. However, if `solidus_auth_devise` is
installed, that is changed after most things have been configured.

It need not be this way: If we want this gem to set a default, we should
set it really early on, not in one of the last initializers to run (they
also run alphabetically, so `solidus_auth_devise` initializers run quite
late.