Chore(config): Stop respecting old configuration
`Spree::Config#redirect_back_on_unauthorized` has been removed in
Solidus 3.0, and we're on 4.4 now. This can all go.

If users have followed the deprecation warnings, then they should be
fine with using `redirect_back`.

Note that we're not using redirect_back in the admin, as logging out
redirects to the admin root controller, which in turn redirects back,
leading to an infinite redirect. Infinite redirects are no good.